+++ tests/test_api.py
@@ -5,10 +5,8 @@ from requests.exceptions import ConnectionError
 
 from memex.api import app
 from memex.auth import gen_token
-from memex.config import read_config
 
-port = read_config()["DEFAULT"]["API_PORT"]
-url = f"http://localhost:{port}/"
+
 headers = {"content-type": "application/json"}
 
 client = app.test_client()